Delete example account router and cleanup things
The router/account.py had unnecessary imports and an endpoint that seemed to serve only as an example, with a hardcoded success message, so it has been deleted. In the core/settings.py and model/account.py, trailing spaces were removed to improve the code readability. In the main.py, the spaces around the assignment operators have been removed as our style guide prefers no spaces around operators in function arguments.