Add product image upload feature and refactor
Implemented a product image upload feature that restricts file types to jpeg or png and limits the file size to 10MB. Renamed account to product for better clarity of use case. Included python-multipart in the requirements for file uploading capabilities. Refactored the save_file method to be asynchronous and included the product id in the directory structure for organisation and easy retrieval.