Update server host in config
The server host in the core environment configuration 'core/env.ini' file was changed from 'localhost' to '0.0.0.0'. This was done to enhance network accessibility as '0.0.0.0' allows the server to listen to requests from any IP address, whereas 'localhost' restricts the server to only listen to requests from the same machine.