"Add feature to download all images as zip
A new capability for downloading all images of a specific owner as a zipfile has been added. For this, the 'get_all_files' function was introduced in 'image_io.py', which archives all the images in a zipfile and returns it as a streaming response. The endpoint for this feature was added in 'image.py' under the route '/get_all_images/{owner_id}/all', which uses the aforementioned function. This feature was required to let users bulk download all of their images at once."