Refactor authentication and account management modules
The update is majorly on streamlining the login routes for clean URLs and enhancing the security of the authentication process. The auth.py utility file was revamped so that authenticate_user now returns TokenData instead of boolean for effective token handling. New methods for user account retrieval and user verification have also been integrated which respectively return user profile and confirm user authentication. The date formatting on the retrieval of user account details from the database was optimized. These changes collectively bolster security, enhance the accuracy of user data and improve the efficiency of data handling.