Refactor image handling in API router
This commit significantly streamlines how images are handled in the API router. Previously, the logic related to image type validation and existence of owner in database was repeated. Now, separate functions have been created for these, reducing the complexity of the `upload_image` and `get_image` functions. Additionally, the `jsonable_encoder` has been removed as it was adding unnecessary complexity to JSONResponse, with dict proving to be sufficient. This refactor leads to improved readability and maintainability of the code.