Refactor Shop routes and improve error handling
This commit refactor shop and image routes by improving error handling and response creation. The helper function `prepare_response` has been implemented to reduce redundancy in response creation, it will generate the proper response based on the result of the operations. A check is also added in `create_shop` route to assert that a shop does not already exist for a given account, this to prevent duplicate shops. The use of HTTPException has been introduced in image routes to provide detailed error messages when exceptional situations like non-existence of a shop for a given account. This refactor has improved the satiability and readability of the code.