Add owner retrieval function in image handling
Imported Account in image_io.py and created a new owner retrieval function `get_owner_uuid`. This was done to allow us to flexibly determine the owner id based on the availability of shop or product uuid. Changes were applied in router/image.py to utilize the new function in both upload_image and get_image routes. This allows image handling to be specific for a shop or a product, improving the system's scalability and flexibility.